requirements_full.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
av==16.0.1
blinker==1.9.0
ccimport==0.4.4
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cumm-cu118==0.7.11
cycler==0.12.1
Cython==3.1.2
dash==3.2.0
dataclasses-json==0.6.7
decorator==5.2.1
decord==0.6.0
Deprecated==1.3.1
diff_gaussian_rasterization @ git+https://github.com/graphdeco-inria/diff-gaussian-rasterization.git@59f5f77e3ddbac3ed9db93ec2cfe99ed6c5d121d
dill==0.4.0
distro==1.9.0
easydict==1.13
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
fire==0.7.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.9.post1/flash_attn-2.5.9.post1+cu118torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=0b6ddd4acf99938100ca53764c98b839364f4473efddadc057649dcbbe9cc236
Flask==3.1.2
fonttools==4.59.0
fsspec==2025.3.2
git-filter-repo==2.47.0
glcontext==3.0.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipycanvas==0.14.1
ipyevents==2.0.4
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==7.4.9
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kaolin==0.18.0
kiwisolver==1.4.8
lark==1.3.1
lpips==0.1.4
Markdown==3.10
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.7.5
matplotlib-inline==0.2.1
moderngl==5.12.0
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numpy==1.26.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
open3d==0.18.0
openai==1.106.1
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
parso==0.8.5
pccm==0.4.16
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
plotly==6.4.0
pluggy==1.6.0
polyscope==2.5.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==3.0.1
pybullet==3.2.7
pydantic==2.12.4
pydantic_core==2.41.5
pygltflib==1.16.5
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.11
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
retrying==1.4.2
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
spconv-cu118==2.3.8
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.4.0+cu118
torch_tensorrt==2.4.0+cu118
torchvision==0.19.0+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trimesh==4.8.1
triton==3.0.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
usd-core==25.5
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
vox2seq @ file:///home/rdagli/code/vomp/extensions/vox2seq
warp-lang==1.10.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==2.0.0
xformers==0.0.27.post2+cu118
zipp==3.23.0
zstandard==0.23.0
